requirements.txt
google-cloud-vision
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
requests
rich
pillow